- dercrease enemies intangible time after hit to 200ms - block that can be broken by shooting or jumping on it - fix projectile trajectories up and down - fix replace itself position # EDDIE:     - npc amigável via behavior flag     - nova IA: throw item     - nova IA: teleport out - EDITOR: shuld not be able to select empty tile - map gfx animated bg: splits image into half for frames - fix jump limits when inside quicksand - new IA para walk ou fly: up to ground, down to inside ground - corrigido projetil tipo TARGET_EXACT - suporte a goto nas reactions - ao trocar de mapa, só deve manter o scroll se o fundo for o mesmo, senao deve resetar - EDITOR: ask if user wants to same data before loading another game - EDITOR: tornar mais fácild e enxergar linhas de tamanho de sala e seleção de tile - active platform e destructible item (sem drop), fora da tela, tem que resetar posição e estado - ao transicionar tela na horizontal e não fort porta, não mudar para STAND se estiver em slide - use hitbox area defined in editor for players instead of hardcoded one - se usar o frog para sair da tela por cima, acaba ficando preso fora dela - novo tipo de escudo: usando escudo (anim_type_shield) - novo tipo de escudo: back - limit the max of spawned npcs from the origin point of the map-npc instead of the number of the main-npc data. - two new object types, destructible (with regular weapons) and same that drops item. - try to use ANIM_TYPE_STAND for hitbox if current anim-type is not used, avoiding falling to default frameSize.     # fix collision against npcs with background. - NPC death explosion being shown at start point when falling in hold/ground (should have no explision at all) - fly/to-player esta deixando rastro como dash, não é para ter. - HIGH: reflected projectiles were not being consumed/finished. - AI: use move-speed to determine the range of the jump. - new AI: walk ahead even if falls from platform. - new AI: shoot projectile in several directions at same time. - new AI: fall down in diagonal. - new projectile types that fire multiple, 5 way up, 5 way down and 8 way. - EDITOR: opção "draw areas" que mostra o retangulo de objetos e inimigos. - moving platform up and down. - EDITOR: add option for FG auto-scroll. - EDITOR: npc with background was not being correctly shown when facing right direction. - EDITOR: poder ver o preview em outros idiomas, respeitando a opção do combo em - EDITOR: adicionada opção view->tileset. - EDITOR: reset do mapa deve ser geral, incluindo links, inimigos e objetos. - EFEITO wave no fundo (bubbleman wily wars).

+++ character/movement/jump.h
@@ -32,10 +32,10 @@ private:
     float speed;
     E_JUMP_STATE state;
     short jumps_number;               // how many consecutive jumps used (for double/triple jumping)
-    short moved;                        // this is used to hardcode the number of pixels a jump should move in maximum
+    float moved;                        // this is used to hardcode the number of pixels a jump should move in maximum
     bool is_bigjump;
     double JUMP_ACCELERATION;
-    int JUMP_LIMIT;
+    float JUMP_LIMIT;
     int start_terrain_type;
     bool start_big_jump;
     double JUMP_INITIAL_SPEED;
